パースとJWW_TEMPを中心部から分離
