+++ src/lib.rs
@@ -1,22 +1,20 @@
+mod io;
 mod transition;
-mod jww;
 
-use anyhow::Result;
-use jww::{
-    jww_temp::JwwTemp,
-    param::{Func, Param},
-};
+use anyhow::{Result, Ok};
+use io::args::Args;
 use std::ffi::OsString;
 
 /// 配線する
 pub fn layout(args: impl IntoIterator<Item = OsString>) -> Result<()> {
-    let param = Param::parse(args)?;
+    let args = Args::parse(args)?;
 
-    let mut jww_temp = JwwTemp::new(&param.file);
+    // let mut jww_temp = JwwTemp::new(&args.file);
 
-    match param.func {
-        Func::Tc(param) => transition::draw(&mut jww_temp, &param),
-    }
+    // match args.func {
+    //     // Func::Tc(param) => transition::draw(&param),
+    // }
 
-    jww_temp.save()
+    // jww_temp.save()
+    Ok(())
 }
