+++ src/transition/mod.rs
@@ -1,7 +1,11 @@
-use crate::jww::{jww_temp::JwwTemp, param::Tc};
-use anyhow::Result;
+pub mod param;
+pub mod polyline;
+
+use anyhow::{Result, bail};
+use self::param::Param;
 
 /// 緩和曲線を描画する
-pub fn draw(jww_temp: &mut JwwTemp, param: &Result<Tc>) {
+pub fn draw(param: &Param) -> Result<polyline::Polyline> {
     // plot(jww_temp, param);
+    bail!("未実装")
 }
