+++ src/transition/polyline.rs
@@ -1,0 +1,4 @@
+/// 緩和曲線ポリライン
+pub struct Polyline {
+
+}
